requirements.txt
streamlit
streamlit-option-menu
audio_recorder_streamlit
wikipedia
textwrap3
spacy
https://huggingface.co/spacy/en_core_web_sm/resolve/main/en_core_web_sm-any-py3-none-any.whl
pillow
IPython
pytesseract
